Move main to __init__ and enable `python -m jockey`.
